server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "htmlmin>=0.1.12",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.18",
    "orjson>=3.11.3",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.2",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn>=0.35.0",
]
